% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flags.R
\name{geom_flag_uf}
\alias{geom_flag_uf}
\title{Get path to state flag}
\usage{
geom_flag_uf(mapping = NULL, data = NULL, size = 0.05, ...)
}
\arguments{
\item{uf}{State abbreviation (e.g. "rj")}
}
\value{
File path to flag image
}
\description{
Get path to state flag
}
