% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flag_circle_uf.R
\name{get_flag_circle_uf}
\alias{get_flag_circle_uf}
\title{Add Brazilian state flags to ggplot}
\usage{
get_flag_circle_uf(uf)
}
\arguments{
\item{uf}{State abbreviation vector (e.g. c("RJ", "SP"))}
}
\value{
Character vector with file paths
}
\description{
Add Brazilian state flags to ggplot
}
